v43_3_staging/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.2
